Reorganize the protocol sources a bit
