+++ communication/src/coap_channel.cpp
@@ -20,7 +20,7 @@
 #undef LOG_COMPILE_TIME_LEVEL
 
 #include "coap_channel.h"
-#include "coap_channel_new.h"
+#include "v2/coap_channel.h"
 #include "service_debug.h"
 #include "messages.h"
 #include "communication_diagnostic.h"
@@ -68,7 +68,7 @@ void CoAPMessageStore::message_timeout(CoAPMessage& msg, Channel& channel)
 		g_unacknowledgedMessageCounter++;
 		// XXX: This will cancel _all_ messages with a timeout error, not just the timed out one.
 		// That's not ideal but should be okay while we're transitioning to the new CoAP API
-		experimental::CoapChannel::instance()->close(SYSTEM_ERROR_COAP_TIMEOUT);
+		v2::CoapChannel::instance()->close(SYSTEM_ERROR_COAP_TIMEOUT);
 		channel.command(MessageChannel::CLOSE);
 	}
 }
@@ -146,7 +146,7 @@ ProtocolError CoAPMessageStore::receive(Message& msg, Channel& channel, system_t
 		}
 		if (msgtype==CoAPType::RESET) {
 			LOG(WARN, "Received RST message; discarding session");
-			experimental::CoapChannel::instance()->handleRst(msg);
+			v2::CoapChannel::instance()->handleRst(msg);
 			if (coap_msg) {
 				coap_msg->notify_delivered_nak();
 			}
